Divided all as components
