+++ card-widget/index.html
@@ -7,6 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>🃏 Card Widgets</title>
     <!-- your styles -->
+    <link rel="stylesheet" href="../shared/style.css" />
     <link rel="stylesheet" href="widget.css" />
     <!-- Roboto Slab for card faces -->
     <link href="https://fonts.googleapis.com/css2?family=Roboto+Slab&display=swap" rel="stylesheet" />
@@ -34,7 +35,7 @@
     <div id="preview"></div>
 
     <!-- your functionality -->
-    <script src="widget.js"></script>
+    <script type="module" src="widget.js"></script>
 </body>
 
 </html>
