--- card-widget/style.css
@@ -1,208 +1,0 @@
-/* Reset and Base Styling */
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  font-family: 'Roboto Slab', serif;
-  background: url("pirate-bg.png") no-repeat left center fixed;
-  background-size: cover;
-  color: #f4f4f4;
-  min-height: 100vh;
-  text-align: center;
-  padding: 2rem;
-  position: relative;
-  backdrop-filter: brightness(0.85);
-}
-
-body::before {
-  content: "";
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 100%;
-  background: rgba(0, 0, 0, 0.5);
-  pointer-events: none;
-  z-index: -1;
-}
-
-/* Header */
-header h1 {
-  font-size: 3rem;
-  margin-bottom: 0.5rem;
-  color: #ffda77;
-  text-shadow: 2px 2px #000;
-}
-
-header p {
-  font-size: 1.2rem;
-  margin-bottom: 2rem;
-  color: #ddd;
-}
-
-/* Buttons */
-.buttons {
-  margin-bottom: 1.5rem;
-}
-
-button {
-  background: linear-gradient(to right, #ffcc70, #ffae42);
-  border: none;
-  padding: 0.7rem 1.5rem;
-  margin: 0 0.5rem;
-  font-size: 1rem;
-  border-radius: 12px;
-  color: #000;
-  font-weight: bold;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.4);
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-button:hover {
-  transform: translateY(-3px);
-  background: linear-gradient(to right, #ffd700, #ff8c00);
-}
-
-/* Card Count */
-.count {
-  font-size: 1.3rem;
-  margin: 1rem 0;
-  font-weight: bold;
-  color: #eee;
-}
-
-/* Card Display Area */
-#preview {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 1rem;
-  padding: 2rem 0;
-}
-
-/* Card Container */
-.card {
-  width: 120px;
-  height: 170px;
-  perspective: 1000px;
-  cursor: pointer;
-}
-
-.card-inner {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  transition: transform 0.8s;
-  transform-style: preserve-3d;
-}
-
-.card.flipped .card-inner {
-  transform: rotateY(180deg);
-}
-
-/* Card Faces */
-.card-front,
-.card-back {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  backface-visibility: hidden;
-  border-radius: 12px;
-  box-shadow: 0 6px 12px rgba(0, 0, 0, 0.4);
-  font-family: 'Roboto Slab', serif;
-}
-
-/* Card Front */
-.card-front.red {
-  background-color: #f4e8d4;
-  color: #b30000;
-  position: relative;
-}
-
-.card-front.black {
-  background-color: #f4e8d4;
-  color: black;
-  position: relative;
-}
-
-/* Card Back */
-.card-back {
-
-  color: #fff;
-  transform: rotateY(180deg);
-  font-size: 2rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  background-color: #ebe6de;
-  background-image:
-    repeating-linear-gradient(45deg,
-      #be1c1c,
-      #be1c1c 1.7px,
-      transparent 1.7px,
-      transparent 6px),
-    repeating-linear-gradient(-45deg,
-      #be1c1c,
-      #be1c1c 1.7px,
-      transparent 1.7px,
-      transparent 6px);
-  border-radius: 10px;
-  box-shadow:
-    inset 0 0 0 6px #f4e8d4,
-    0 2px 6px rgba(0, 0, 0, 0.2),
-    0 4px 12px rgba(0, 0, 0, 0.15);
-}
-
-/* Card Corners */
-.card-corner {
-  position: absolute;
-  font-size: 1rem;
-  line-height: 1.1;
-  text-align: center;
-  font-weight: bold;
-}
-
-.top-left {
-  top: 6px;
-  left: 6px;
-}
-
-.bottom-right {
-  bottom: 6px;
-  right: 6px;
-  transform: rotate(180deg);
-}
-
-
-
-.card-center {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  font-size: 2.5rem;
-  pointer-events: none;
-  user-select: none;
-}
-
-.face-card {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  text-align: center;
-}
-
-.face-card img {
-  width: 70px;
-  height: auto;
-  display: block;
-  margin: 0 auto;
-}
